core: Add initialize_dependencies
This function is responsible to traverse all dependencies of bears
correctly and instantiate those using a consumer-based system (so a
dependency bear is not run more than once, limited to section +
file-dictionary). It also sets up a DependencyTracker instance to track
all bear instance and their dependencies/dependants, and it also
returns a set of bears that have no dependencies and can be run.

Dependencies are not yet integrated into the core.